Crate sandbox

Módulo principal, Lógica de execução, renderização, inicialização e chamada às funções.

Bibliotecas e extras

Para realização do projeto, utilizamos a biblioteca Pixels https://docs.rs/pixels/latest/pixels/ para abstrair a necessidade de criação de shaders com o wgpu, permitindo uma renderização mais simplificada. Para o Handling de entrada e criação de janelas, foi utilizada a biblioteca winit, https://docs.rs/winit/latest/winit/ , que permite tratar input e realizar a criação de janelas mais diretamente.

Além disso, para o processo de documentação e criação dessa wiki, foi utilizada a feature do cargo docs, o qual permite uma criação de documentação próxima a do Doxygen.

Modules

implparticles 🔒

Implementação das traits base para cada partícula e definição de position_to_index

particle 🔒

Definição das Structs para cada Partícula, declaração das traits básicas de movimento e colisão, definição de Enums para os tipos das partículas.

Functions

draw

Renderização

instanceparticle

Instanciação de Partículas

main 🔒

Execução Prinicipal

update

Atualização de Partículas