Crate sandbox

Basic Pod Sandbox types

Modules

error

pinned

A pod sandbox implementation which does pin it’s namespaces to file descriptors.

pinns

Structs

LinuxNamespaces

Sandbox

This is the main data structure for a Pod Sandbox. The implementation T can vary and is being defined in the Pod trait. Responsibility of the Sandbox is to hold arbitrary necessary data for the implementation and not modify it in any way.

SandboxBuilder

Builder for Sandbox.

SandboxConfig

SandboxData holds all the data which will be passed around to the Pod trait, too.

SandboxConfigBuilder

Builder for SandboxConfig.

SandboxContext

SandboxContextBuilder

Builder for SandboxContext.

SandboxState

SecurityConfig

SecurityConfigBuilder

Builder for SecurityConfig.

Traits

Pod